refactor: move ServicesSection to ui folder and keep ServicesCard in the page component and importation srvicesCardHome Component in the home page
