+++ src/app/page.jsx
@@ -3,12 +3,13 @@ import ValueSection from "@/components/pages/home/sections/ValueSection/ValueSec
 import EssentialsSection from "@/components/pages/home/sections/essentialsSection/EssentialsSection";
 import HistorySection from "@/components/pages/home/sections/historySection/HistorySection";
 import Testimonials from "@/components/pages/home/sections/testimonial/Testimonial";
-import ServicesSection from "@/components/pages/home/sections/servicesSection/ServicesSection";
+import ServicesSection from "@/components/ui/services/ServicesSection";
 import PartnerSection from "@/components/pages/home/sections/PartnersSection/PartnersSection";
 import PhilosophySection from "@/components/pages/home/sections/philosophySection/PhilosophySection";
 import RealizationsSection from "@/components/pages/home/sections/realizationSection/RealizationSection";
 import NewsletterSignup from "@/components/pages/home/sections/newLetter/NewLetter";
 import { carouselSlides } from "@/components/pages/home/data/carouselSlides";
+import ServicesCardHome from "@/components/pages/home/sections/servicesSection/ServicesCardHome";
 
 
 export default function Home() {
@@ -24,7 +25,7 @@ export default function Home() {
       <EssentialsSection />
       <ValueSection />
       <HistorySection />
-      <ServicesSection />
+      <ServicesSection ServicesCard={ServicesCardHome} />
       <RealizationsSection />
       <Testimonials />
       <PhilosophySection />
